requirements.txt
transformers==4.26.1
torch==1.8.0
torchvision==0.9.0
torchtext==0.10.0
Pillow==9.4.0
opencv-python-headless==4.1.2.30

-r lama/requirements.txt